feat(settings): 新增设置模块及界面组件
新增设置模块，包括设置窗口、界面组件和工具函数。设置模块支持学生编辑、分组编辑、界面设置、关于和更新等功能。新增了颜色选择器、滑块组件、分组卡片、分组编辑框和更新确认框等组件，并提供了工具函数用于窗口大小计算、颜色判断和配置管理。
