+++ settings/components/__init__.py
@@ -1,0 +1,14 @@
+"""RandPicker 设置组件模块。
+
+此模块包含所有设置界面使用的组件实现。
+"""
+
+from .slider_component import create_weight_slider_widget
+from .color_picker import create_color_dialog, update_color_preview
+from .table_widget import GroupCard, GroupEditBox, GroupEnablePolicyBox, UpdateConfirmBox
+
+__all__ = [
+    'create_weight_slider_widget',
+    'create_color_dialog', 'update_color_preview',
+    'GroupCard', 'GroupEditBox', 'GroupEnablePolicyBox', 'UpdateConfirmBox'
+]
